libs/python/shared/py-debug/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "pydebug" ]

[tool.coverage.report]
exclude_lines = [ "if TYPE_CHECKING:" ]
show_missing = true

[tool.pytest.ini_options]
addopts = "--cov --cov-fail-under=80 --cov-report html:'../../../../coverage/libs/python/shared/py-debug/html' --cov-report xml:'../../../../coverage/libs/python/shared/py-debug/coverage.xml' --html='../../../../reports/libs/python/shared/py-debug/unittests/html/index.html' --junitxml='../../../../reports/libs/python/shared/py-debug/unittests/junit.xml'"

[tool.poetry]
name = "python-shared-py-debug"
version = "1.0.0"
description = "Automatically generated by Nx."
authors = [ ]
license = "Proprietary"
readme = "README.md"

  [[tool.poetry.packages]]
  include = "pydebug"

  [tool.poetry.dependencies]
  python = "^3.10"

    [tool.poetry.dependencies.shared-py-log]
    path = "../py-log"
    develop = true

[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"
